modules/music-importer/requirements.txt
patool~=1.12
python-dotenv~=1.0.0
mutagen~=1.47.0
librosa~=0.10.2.post1
pymysql~=1.1.1
requests~=2.32.4
yt-dlp~=2025.7.21
coverage~=7.2.7
telethon~=1.32.0
fastapi~=0.111.0
uvicorn~=0.22.0

anyio~=4.10.0
numpy~=2.0.2